requirements.txt
gpiozero~=1.6.2
wiringpi
python-dotenv
paho-mqtt
